Changing name member variable to std::string to make python binding easier
When pybind11 passes a python string to a C++ function expecting a char*
it seems to create temporary memory that exists for the duration of the function
call. It appears the SDK doing a simple copy of that for the constructor
causes it to fail in the python bindings. Switching to a std::string to
force a copy